api for question and answer

+++ app/api/user.py
@@ -2,28 +2,39 @@ from flask import jsonify, request, current_app, url_for
 from . import api
 from app.models import User
 from app import db
+from app.util import json_data
 
 
 @api.route('/user/<int:id>')
 def get_user(id):
-    user = User.query.get_or_404(id)
-    return jsonify(user.to_json())
+    user = User.query.get(id)
+    if user is None:
+        return jsonify({'code': 0, 'msg': 'user not exist'})
+    else:
+        return jsonify({'code': 1, 'data': user.to_json()})
 
 
 @api.route('/user')
-def get_user_by_username():
+def get_user_by_args():
+    if not request.args:
+        users = User.query.all()
+        if not users:
+            return jsonify({'code': 0, 'msg': 'no data'})
+        else:
+            return jsonify({'code': 1, 'data': [user.to_json() for user in users]})
+
     username = request.args.get('username')
-    user = User.query.filter(User.user_username == username).first_or_404()
-    return jsonify(user.to_json())
+    if username:
+        user = User.query.filter(User.user_username == username).first()
+        if user is None:
+            return jsonify({'code': 0, 'msg': 'user not exist'})
+        else:
+            return jsonify({'code': 1, 'data': user.to_json()})
+    else:
+        return json_data(0, 'bad request')
 
 
-@api.route('/user/')
-def get_users():
-    users = User.query.all()
-    return jsonify([user.to_json() for user in users])
-
-
-@api.route('/user/', methods=['POST'])
+@api.route('/user', methods=['POST'])
 def add_user():
     json = request.json
     user_username = json.get('user_username')
@@ -34,15 +45,19 @@ def add_user():
     if user_username is None or user_username == '' \
             or user_password is None or user_password == '' \
             or user_type is None or user_type == '':
-        return jsonify({'error': 'bad data'}), 400
+        return jsonify({'code': 0, 'msg': 'bad request'})
 
     if User.query.filter(User.user_username == request.json.get('user_username')).all():
-        return jsonify({'error': 'username existed'}), 400
+        return jsonify({'code': 0, 'msg': 'username existed'})
 
     user = User(user_username, user_password, user_type, user_authentication)
     db.session.add(user)
-    db.session.commit()
-    return jsonify(user.to_json()), 201
+    try:
+        db.session.commit()
+    except Exception, e:
+        db.session.rollback()
+        return json_data(0, e.message)
+    return jsonify({'code': 1, 'data': user.to_json()})
 
 
 @api.route('/user/<int:id>', methods=['PUT'])
@@ -53,9 +68,11 @@ def update_user(id):
     user_authentication = json.get('user_authentication', None)
 
     if user_password == '' or user_type == '':
-        return jsonify({'error': 'bad data'}), 400
+        return jsonify({'code': 0, 'msg': 'bad request'})
 
-    user = User.query.get_or_404(id)
+    user = User.query.get(id)
+    if user is None:
+        return jsonify({'code': 0, 'msg': 'user not exist'})
 
     if user_password:
         user.user_password = user_password
@@ -63,15 +80,24 @@ def update_user(id):
         user.user_type = user_type
     if user_authentication:
         user.user_authentication = user_authentication
-
-    db.session.commit()
-    return jsonify(user.to_json()), 201
+    try:
+        db.session.commit()
+    except Exception, e:
+        db.session.rollback()
+        return json_data(0, e.message)
+    return jsonify({'code': 1, 'data': user.to_json()})
 
 
 @api.route('/user/<int:id>', methods=['DELETE'])
 def delete_user(id):
-    user = User.query.get_or_404(id)
+    user = User.query.get(id)
+    if user is None:
+        return jsonify({'code': 0, 'msg': 'user not exist'})
 
     db.session.delete(user)
-    db.session.commit()
-    return jsonify(None), 204
+    try:
+        db.session.commit()
+    except Exception, e:
+        db.session.rollback()
+        return json_data(0, e.message)
+    return jsonify({'code': 1})
